isReachable/isPaired as properties, correct icon
This change switches is_reachable and is_paired to fetch the properties
instead of trying to call a method.  (not sure if/when this changed in
the dbus interface, but that's how they are presented for me)

This fixes issue #31 and (I think) #38

This also switches the icon code to use the statusIconName as that seems
to have the correct value.  Also, in DeviceIndicator, no longer append
"-symbolic" as (at least on my system) the icons don't seem to have
that in their names.